工程实现了点到360条线的计算

+++ simulate/inc/read.h
@@ -2,10 +2,10 @@
 #define __READ_H
 typedef struct {
     long long lat;
-    long double latd;
+    double latd;
     char lat_dir;
     long long lon;
-    long double lond;
+    double lond;
     char lon_dir;
 }  GPS_msgTypeDef;
